pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "kontiki-tui"
version = "0.1.0"
description = "Terminal UI to monitor Kontiki services, registry events, and logs."
authors = ["The Kontiki Authors"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "kontiki_tui"}]
include = [
    {path = "kontiki_tui/app.css", format = ["sdist", "wheel"]},
    {path = "LICENSE", format = ["sdist", "wheel"]},
    {path = "NOTICE", format = ["sdist", "wheel"]},
]
keywords = ["kontiki", "tui", "textual", "monitoring", "amqp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console :: Curses",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
textual = "*"
psutil = "^7.2.2"
pyyaml = "^6.0.3"
rich = "^14.3.3"
kontiki = ">=1.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^7.0.0"

[tool.poetry.scripts]
kontiki-tui = "kontiki_tui.app:main"

[tool.poetry.urls]
Homepage = "https://github.com/kontiki-org/kontiki-tui"
Repository = "https://github.com/kontiki-org/kontiki-tui"
Issues = "https://github.com/kontiki-org/kontiki-tui/issues"


[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["kontiki_tui"]
